--- a/violet-backend-core/Cargo.toml
+++ b/violet-backend-core/Cargo.toml
@@ -9,3 +9,4 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1.0", features = ["full"] }
 redis = "0.23"
+tempfile = "3.8"