… violet-backend-core/Cargo.lock
… violet-backend-core/src/main.rs
… violet-backend-core/src/memory.rs