- top access

```
Starting stress test with 50000000 entries...

Insertion completed in 135.8405554s
Insertion rate: 368078.59 entries/sec

zrevrange query completed in 8.767µs (average over 100 iterations)
zrevrange query rate: 114064.10 queries/sec

Total entries in memory: 50000000
Memory usage per RankedEntry: ~32 bytes
Memory usage per ExpireEntry: ~32 bytes
Memory optimization: Original ExpireEntry size: ~48 bytes
Memory optimization: Current ExpireEntry size: ~32 bytes
Memory optimization: Saving ~16 bytes per entry
Total memory saved: ~762 MB
Total entries in expire queue: 50000000
Number of unique tables: 1
Memory usage for table management: ~0 KB
```

- random access

```
Starting stress test with 50000000 entries...

Insertion completed in 89.5489589s
Insertion rate: 558353.78 entries/sec

zrange query completed in 42.86083ms (average over 100 iterations)
zrange query rate: 23.33 queries/sec

Total entries in memory: 50000000
Memory usage per RankedEntry: ~32 bytes
Memory usage per ExpireEntry: ~32 bytes
Memory optimization: Original ExpireEntry size: ~48 bytes
Memory optimization: Current ExpireEntry size: ~32 bytes
Memory optimization: Saving ~16 bytes per entry
Total memory saved: ~762 MB
Total entries in expire queue: 50000000
Number of unique tables: 1
Memory usage for table management: ~0 KB
```

- with paging
```
Starting page stress test with 50000000 entries...

Insertion completed in 109.3602173s
Insertion rate: 457204.65 entries/sec

Total pages created: 0

zrevrange query completed in 8.899µs (average over 100 iterations)
zrevrange query rate: 112372.18 queries/sec
Current entries in memory: 50000000
Memory usage per entry: ~32 bytes
Total page file size: 0 MB
```