@violet-dev 2025-05-03
```rs
use std::{cmp::Ordering, ops::Range};

#[derive(Debug)]
struct Node<T> {
    val: T,
    left: Option<Box<Node<T>>>,
    right: Option<Box<Node<T>>>,
    height: usize,
    size: usize, // 왼쪽 + 오른쪽 + 자기자신 개수
}

impl<T: Ord> Node<T> {
    fn new(val: T) -> Self {
        Node {
            val,
            left: None,
            right: None,
            height: 1,
            size: 1,
        }
    }
}

#[derive(Debug)]
pub struct IndexableSet<T: Ord> {
    root: Option<Box<Node<T>>>,
}

impl<T: Ord> IndexableSet<T> {
    pub fn new() -> Self {
        Self { root: None }
    }

    pub fn insert(&mut self, val: T) {
        self.root = Self::insert_inner(self.root.take(), val);
    }

    pub fn get(&self, index: usize) -> Option<&T> {
        Self::get_inner(&self.root, index)
    }

    pub fn range(&self, range: Range<usize>) -> Vec<&T> {
        let mut result = Vec::with_capacity(range.end.saturating_sub(range.start));
        Self::range_inner(&self.root, range, &mut 0, &mut result);
        result
    }

    pub fn take(&mut self, val: &T) -> Option<T> {
        let (new_root, removed) = Self::take_inner(self.root.take(), val);
        self.root = new_root;
        removed
    }

    fn height(n: &Option<Box<Node<T>>>) -> usize {
        n.as_ref().map_or(0, |node| node.height)
    }

    fn size(n: &Option<Box<Node<T>>>) -> usize {
        n.as_ref().map_or(0, |node| node.size)
    }

    fn update(node: &mut Box<Node<T>>) {
        node.height = 1 + Self::height(&node.left).max(Self::height(&node.right));
        node.size = 1 + Self::size(&node.left) + Self::size(&node.right);
    }

    fn get_inner<'a>(node: &'a Option<Box<Node<T>>>, index: usize) -> Option<&'a T> {
        let node = node.as_ref()?;
        let lsize = Self::size(&node.left);
        if index < lsize {
            Self::get_inner(&node.left, index)
        } else if index == lsize {
            Some(&node.val)
        } else {
            Self::get_inner(&node.right, index - lsize - 1)
        }
    }

    fn insert_inner(node: Option<Box<Node<T>>>, val: T) -> Option<Box<Node<T>>> {
        match node {
            Some(mut node) => {
                match val.cmp(&node.val) {
                    Ordering::Less => node.left = Self::insert_inner(node.left, val),
                    Ordering::Greater => node.right = Self::insert_inner(node.right, val),
                    Ordering::Equal => return Some(node), // 이미 존재
                }
                Self::update(&mut node);
                Some(node) // (AVL 회전 생략: 간단 버전)
            }
            None => Some(Box::new(Node::new(val))),
        }
    }

    fn range_inner<'a>(
        node: &'a Option<Box<Node<T>>>,
        range: Range<usize>,
        cur_index: &mut usize,
        output: &mut Vec<&'a T>,
    ) {
        if let Some(ref n) = node {
            // 왼쪽부터 순회
            Self::range_inner(&n.left, range.clone(), cur_index, output);

            if range.contains(cur_index) {
                output.push(&n.val);
            }

            *cur_index += 1;

            if *cur_index >= range.end {
                return; // 종료 조기 탈출
            }

            Self::range_inner(&n.right, range, cur_index, output);
        }
    }

    fn take_inner(node: Option<Box<Node<T>>>, val: &T) -> (Option<Box<Node<T>>>, Option<T>) {
        match node {
            Some(mut n) => match val.cmp(&n.val) {
                Ordering::Less => {
                    let (new_left, removed) = Self::take_inner(n.left, val);
                    n.left = new_left;
                    Self::update(&mut n);
                    (Some(n), removed)
                }
                Ordering::Greater => {
                    let (new_right, removed) = Self::take_inner(n.right, val);
                    n.right = new_right;
                    Self::update(&mut n);
                    (Some(n), removed)
                }
                Ordering::Equal => {
                    let removed_val = n.val;

                    match (n.left, n.right) {
                        (None, None) => (None, Some(removed_val)),
                        (Some(left), None) => (Some(left), Some(removed_val)),
                        (None, Some(right)) => (Some(right), Some(removed_val)),
                        (Some(left), Some(right)) => {
                            let (new_right, min_node_opt) = Self::take_min(right);
                            if let Some(min_node) = min_node_opt {
                                let mut new_node = Box::new(Node {
                                    val: min_node.val,
                                    left: Some(left),
                                    right: new_right,
                                    height: 0,
                                    size: 0,
                                });
                                Self::update(&mut new_node);
                                (Some(new_node), Some(removed_val))
                            } else {
                                unreachable!("right subtree had at least one node, so take_min must return Some");
                            }
                        }
                    }
                }
            },
            None => (None, None),
        }
    }

    fn take_min(mut node: Box<Node<T>>) -> (Option<Box<Node<T>>>, Option<Box<Node<T>>>) {
        match node.left {
            Some(left_node) => {
                let (new_left, min_node) = Self::take_min(left_node);
                node.left = new_left;
                Self::update(&mut node);
                (Some(node), min_node)
            }
            None => {
                // 현재 노드가 최소 노드
                let right = node.right.take();
                (right, Some(node))
            }
        }
    }
}


````